botlib: C++ compatibility for be_aas_move

+++ code/botlib/be_aas_move.c
@@ -168,7 +168,7 @@ int AAS_AgainstLadder(vec3_t origin)
 		//get the plane the face is in
 		plane = &aasworld.planes[face->planenum ^ side];
 		//if the origin is pretty close to the plane
-		if (fabs(DotProduct(plane->normal, origin) - plane->dist) < 3)
+		if (fabs((double)(DotProduct(plane->normal, origin) - plane->dist)) < 3)
 		{
 			if (AAS_PointInsideFace(abs(facenum), origin, 0.1f)) return qtrue;
 		} //end if
@@ -394,7 +394,7 @@ static void AAS_ApplyFriction(vec3_t vel, float friction, float stopspeed,
 	float speed, control, newspeed;
 
 	//horizontal speed
-	speed = sqrt(vel[0] * vel[0] + vel[1] * vel[1]);
+	speed = sqrt((double)(vel[0] * vel[0] + vel[1] * vel[1]));
 	if (speed)
 	{
 		control = speed < stopspeed ? stopspeed : speed;
@@ -1084,7 +1084,7 @@ int AAS_HorizontalVelocityForJump(float zvel, vec3_t start, vec3_t end, float *v
 		return 0;
 	}
 	//calculate horizontal speed
-	*velocity = sqrt(dir[0]*dir[0] + dir[1]*dir[1]) / (t + zvel / phys_gravity);
+	*velocity = sqrt((double)(dir[0]*dir[0] + dir[1]*dir[1])) / (t + zvel / phys_gravity);
 	//the horizontal speed must be lower than the max speed
 	if (*velocity > phys_maxvelocity)
 	{
